test(auth):complete auth/profile interface
