+++ .gitignore
@@ -35,5 +35,6 @@ build/
 # unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
 .rvmrc
 
-
+.DS_Store
 Gemfile.lock
+/spec/integration/
