+++ lib/alauda/api/application.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Application
+
+    end
+  end
+end
