+++ lib/alauda/api/backup.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Backup
+
+    end
+  end
+end
