+++ lib/alauda/api/hook.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Hook
+
+    end
+  end
+end
