+++ lib/alauda/api/instance.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Instance
+
+    end
+  end
+end
