+++ lib/alauda/api/job.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Job
+
+    end
+  end
+end
