+++ lib/alauda/api/repository.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Repository
+
+    end
+  end
+end
