+++ lib/alauda/api/service.rb
@@ -1,0 +1,7 @@
+module Alauda
+  module Api
+    class Service
+
+    end
+  end
+end
