--- lib/alauda/core.rb
@@ -1,3 +1,0 @@
-#require_relative 'common'
-require_relative 'core/api/application'
-require_relative 'core/api/auth'
