--- lib/alauda/core/api/application.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Application
-
-			end
-		end
-	end
-end
