--- lib/alauda/core/api/auth.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Auth
-
-			end
-		end
-	end
-end
