--- lib/alauda/core/api/backup.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Backup
-
-			end
-		end
-	end
-end
