--- lib/alauda/core/api/hook.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Hook
-
-			end
-		end
-	end
-end
