--- lib/alauda/core/api/instance.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Instance
-
-			end
-		end
-	end
-end
