--- lib/alauda/core/api/job.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Job
-
-			end
-		end
-	end
-end
