--- lib/alauda/core/api/repository.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Repository
-
-			end
-		end
-	end
-end
