--- lib/alauda/core/api/service.rb
@@ -1,9 +1,0 @@
-module Alauda
-	module Core
-		module Api
-			class Service
-
-			end
-		end
-	end
-end
