+++ spec/api/auth_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 describe 'Auth' do
-  describe 'blank' do
-    it 'show Hello Captain' do
-      expect(true).to eq true
+  describe 'profile' do
+    it 'get profile with api key' do
+      
     end
   end
 end
