+++ spec/spec_helper.rb
@@ -1,24 +1,14 @@
 # -*- encoding: utf-8 -*-
 
-require 'rubygems'
-require 'bundler'
-Bundler.setup(:default, :test)
-require 'rspec'
-
-
-
 require "codeclimate-test-reporter"
 ENV['CODECLIMATE_REPO_TOKEN'] = 'db4d1d9caeb8a8828ac467482b85330526ffac8ba268f6ebb900cb672d12709e'
 CodeClimate::TestReporter.start
 
 
-require 'webmock/rspec'
-require 'alauda/core'
-require 'alauda/common'
-WebMock.disable_net_connect!(:allow => "codeclimate.com")
-
+require 'alauda/alauda'
 
 RSpec.configure do |config|
+
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
@@ -27,5 +17,8 @@ RSpec.configure do |config|
     mocks.verify_partial_doubles = true
   end
 
+  config.before(:all)  do 
+    Alauda.api_key = "1000000000000000000000000000000000000000"
+  end
 end
